core: heavy refactor
