+++ backend/src/api/microservices/database/mod.rs
@@ -1,3 +1,2 @@
-pub mod core;
 pub mod init;
 pub mod schema;
