+++ backend/src/api/mod.rs
@@ -1,5 +1,3 @@
 pub mod bot;
 pub mod microservices;
-pub mod models;
-pub mod utilities;
 pub mod web;
