+++ backend/src/api/web/mod.rs
@@ -1,8 +1,9 @@
 pub mod cookies;
 pub mod handlers;
-pub mod home;
-pub mod lock;
+pub mod locks;
+pub mod models;
 pub mod sessions;
 pub mod swap;
 pub mod twofactor;
+pub mod utilities;
 pub mod verify;
