+++ backend/src/api/web/utilities.rs
@@ -1,37 +1,28 @@
-use super::{
-    microservices::redis::try_get,
-    models::{RedisAction, VerifiedTokenResult, WebsitePath},
+use super::models::{RedisAction, VerifiedTokenResult, WebsitePath};
+use crate::{
+    AppError, AppState, WebsiteRoute, api::microservices::redis::try_get, metrics::incr_visitors,
 };
-use crate::{AppError, AppState, WebsiteRoute, metrics::incr_visitors};
 use axum::{extract::Request, http::header::HeaderMap};
 use sha2::{Digest, Sha256};
 use std::{net::IpAddr, sync::Arc};
 
-pub fn get_hashed_ip(headers: &HeaderMap, direct_ip: IpAddr) -> String {
-    let ip = headers
-        .get("cf-connecting-ip")
-        .and_then(|v| v.to_str().ok())
-        .map(|s| s.to_string())
-        .or_else(|| {
-            headers
-                .get("x-forwarded-for")
-                .and_then(|v| v.to_str().ok())
-                .and_then(|s| s.split(',').next().map(|s| s.trim().to_string()))
-        })
-        .unwrap_or_else(|| direct_ip.to_string());
+pub async fn clear_all_keys(
+    state: Arc<AppState>,
+    website_path: &str,
+    keys: &[&str],
+    email: &str,
+) -> Result<(), AppError> {
+    let mut pipe = redis::pipe();
 
-    let mut hasher = Sha256::new();
-    hasher.update(ip.as_bytes());
+    for key in keys {
+        pipe.del(format!("{}:{}:{}", website_path, key, email))
+            .ignore();
+    }
 
-    format!("{:x}", hasher.finalize())
-}
+    pipe.query_async::<()>(&mut state.redis_connection_manager.clone())
+        .await?;
 
-pub fn get_key(redis_action: RedisAction, hashed_ip: &str) -> String {
-    format!("{}:{}", redis_action.as_ref(), hashed_ip)
-}
-
-pub fn convert_i8_to_u8(payload: &i8) -> u8 {
-    payload.checked_abs().unwrap_or(0) as u8
+    Ok(())
 }
 
 pub async fn format_verified_result(
@@ -145,4 +136,27 @@ pub fn get_website_path(label: &str) -> WebsitePath {
             panic!("Added connection invalid")
         }
     }
+}
+
+pub fn get_hashed_ip(headers: &HeaderMap, direct_ip: IpAddr) -> String {
+    let ip = headers
+        .get("cf-connecting-ip")
+        .and_then(|v| v.to_str().ok())
+        .map(|s| s.to_string())
+        .or_else(|| {
+            headers
+                .get("x-forwarded-for")
+                .and_then(|v| v.to_str().ok())
+                .and_then(|s| s.split(',').next().map(|s| s.trim().to_string()))
+        })
+        .unwrap_or_else(|| direct_ip.to_string());
+
+    let mut hasher = Sha256::new();
+    hasher.update(ip.as_bytes());
+
+    format!("{:x}", hasher.finalize())
+}
+
+pub fn get_key(redis_action: RedisAction, hashed_ip: &str) -> String {
+    format!("{}:{}", redis_action.as_ref(), hashed_ip)
 }
